requirements.txt
streamlit
transformers==4.18.0
torch==1.11.0
pdf2image==1.16.0
Pillow==9.1.1
